pyproject.toml
[project]
name = "langchain-study"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langgraph>=0.5.4",
    "langserve[all]>=0.3.1",
    "langsmith>=0.4.8",
    "python-dotenv>=1.1.1",
    "spacy>=3.8.7",
    "xinference-client>=1.8.0",
    "pymilvus>=2.5.14",
    "loguru>=0.7.3",
    "langchain-tavily>=0.2.11",
    "tavily-python>=0.7.11",
    "langgraph-checkpoint-postgres>=2.0.23",
    "psycopg2-binary>=2.9.10",
    "psycopg[binary,pool]>=3.2.9",
    "aiofiles>=24.1.0",
    "daytona>=0.25.6",
    "daytona-sdk>=0.25.6",
    "pytesseract>=0.3.13",
    "chardet==5.2.0",
    "openpyxl==3.1.2",
    "structlog==25.4.0",
    "peewee>=3.18.2",
    "peewee-migrate>=1.13.0",
    "langchain-deepseek>=0.1.4",
]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.1.0",
    "pytest-dotenv>=0.5.2",
    "ruff>=0.12.10",
]
